refact(ServicePage): break down components
